JBIDE-4342: JstWebAllTests/testTlds fails because of webpromptingprovider returning an empty list
It seem that some TestCase from this suite invokes a new job or thread that doesn't allow the TLD-testcase to be succeeded.
The magic with the WaitForIdle() utility and the re-ordering the testcases within the suite are solved the problem on my comptuter. But this is the subject to perform the testing on Hudson.

git-svn-id: https://svn.jboss.org/repos/jbosstools/trunk@15376 a97e2381-89e5-4abb-bab3-167db6db766c