https://issues.jboss.org/browse/JBDS-1876 emma code coverage in overall jbosstools-3.3_trunk.soa-tooling.tests fails to find org.eclipse.bpel.* sources; cannot do coverage report so builds fail
eclemma gather coverage data only for loaded classes, which means final coverage statistic is usually higher that real coverage, because it dosen't consider classes that that wasn't loaded during test. To get real coverage for all classes metadata should be generated manually for plugins included in coverage report. Now it is not possible for plugins that are outside component codebase and are not part of the build. 

I fixed error by skipping metadata generation for unreachable plugins, but it means coverage report shows coverage results higher then they really are.


git-svn-id: https://svn.jboss.org/repos/jbosstools/trunk@37918 a97e2381-89e5-4abb-bab3-167db6db766c